requirements.txt
flask
requests
beautifulsoup4
langchain_openai
langchain_community
python-dotenv